+++ src/tests/test-lightify.c
@@ -485,7 +485,7 @@ START_TEST(lightify_tst_scan_nodes)
 		ck_assert_int_eq(color, 0xf3);
 
 		// Check if firwmare version has been correctly stored.
-		unsigned long version = lightify_node_get_fwversion(node);
+		uint32_t version = lightify_node_get_fwversion(node);
 		ck_assert(version == 0x01020307);
 
 		// Check we delete node information when scanfornodes fails.
